Set package size for connections and send multiple commands
Currently there is only sent 1 pixel per package which is inefficent. It
would be better to set as many pixel as possible per package.

This commit allows to specify a package size and then sends as many
pixels as possible per package.

This is currently not possible to change with an argument, a todo
has been added for that