Make client flush more configurable
Currently the client gets flushed every command (which is every pixel)
this is inefficient since every tcp package can hold multiple pixel (the
exact number depends on the network and mtu).

This patch gives 3 options for setting a FlushMode:
- Manual > Never flush
- Commands > Flush after a number of commands (like Pixels)
- Bytes > Flush when a package size is reached.

Further a flush_size has to be set. Example:

To run in the current mode where a flush occurs per pixel:
- FlushMode::Command,1

To flush at a size of 1500 Bytes (which is a commonly used mtu):
- FlushMode:Bytes,1

Further a method is introduced for the client to flush manually